+++ pre-processing/pre-processing-code/lambda_function.py
@@ -71,7 +71,14 @@ def start_change_set(describe_entity_response, revision_arn):
 def lambda_handler(event, context):
     asset_list = source_dataset(data_set_name, s3_bucket, new_s3_key)
 
-    if type(asset_list) == list and len(asset_list) > 0:
+    if type(asset_list) == list:
+        if len(asset_list) == 0:
+            print(
+                'No need for a revision, all datasets included with this product are up to date')
+            return {
+                'statusCode': 200,
+                'body': json.dumps('No need for a revision, all datasets included with this product are up to date')
+            }
 
         create_revision_response = dataexchange.create_revision(
             DataSetId=data_set_id)